Create a fake framework instead of embedded framework + multi arch build
This probably requires XCode 6 to build.

The problem with embedded frameworks is that they are embedded as is
within the application, so multi archs only wastes space. Moreover,
archs for the simulator makes the app no longer validate in iTunes.
Additionally embedded frameworks implies iOS 8 minimum for the apps.

We basically have to create a framework (for simplicity of use), that
acts like a shared library (for lighter weight in the app).
The modifications are based on:
  https://github.com/kstenerud/iOS-Universal-Framework

The aggregate target builds the framework for both iphoneos and
iphonesimulator, which normally implies: armv7 armv7s arm64 i386,
and x86_64.
Once linked in the application, the added weight is only (roughly)
one fourth (somehow armv7s is not explicitly present) of the whole
framework.