Change permission check to use authorizationStatus instead of alert|sound|badge
On iOS it's possible to accept notifications but have them be totally
invisible to the user: no alert, sound or badge, not on the lock-screen
nor in the notification center.
But once notifications are made visible anew, they are shown as if they
were not gone. That's not the case when switching "Allow Notifications"
off, however.

In both cases iOS shows the app as "Off" in the Notifications settings
screen.
We ignore the first case to be more robust in case of iOS notification
settings change to prevent false negatives. It's also a rather contrived
corner-case so it's better to ignore it.