Make isSubscribedToNotifications default to true
Since v3.1.0 we report the OS notification permission status and flip
the installation as soft opt-out if not granted. However it was still
necessary to call `WonderPush.subscribeToNotifications()` even if your
application handled asking the user for permission itself.

This is the functional equivalent to having two flip switches in series
that are open. Asking the OS for permission left your installation soft
opt-out, and that was unnecessary. What is more, it was a difference
between iOS and Android that is no longer useful.6

Note that `WonderPush.isSubscribedToNotifications()` defaults to `true`
but you still need to call `WonderPush.subscribeToNotifications()` to
ask the user for permission, if you don't handle it yourself.