Improve request vault
I noticed that while requests are being enqueued in the request vault,
and removed if successfully sent, they were not being properly restored
when the request vault was restored.

We were reading the `savedRequests` instance property instead of calling
the `+savedRequests` class selector.

Additionally `NSUnarchiver` was getting confused of the `WPRequest`
class being present both in `WonderPush.framework` and
`WonderPushExtension.framework`
and the `isKindOfClass:` test was failing.

Moreover, we are creating multiple instances of `WPRequestVault` which
were all storing requests under the same `NSUserDefaults` key.

So I implemented the following solution:

1. One `NSUserDefaults` key per instance of `WPRequestVault`.
2. I changed the serialization format to an `NSData` containing the JSON
   serialization of the `WPRequest` class to avoid `NSUnarchiver` from
   potentially picking the `WPRequest` from the wrong source if an
   application is both linked against `WonderPush.framework` and
   `WonderPushExtension.framework`, breaking the usefullness of an
   `isKindOfClass:` test.
3. Proper resumption of the saved requests after `WPRequest` vault
   initialization. I avoided doing so in the initialization because
   `WPAPIClient` creates a `WPRequest` instance during its own
   initialization, giving itself as the `requestExecutor`. So we provide
   means to resume requests manually through the `restoreQueue` method.
4. To avoid `API calls disabled` messages, we call `restoreQueue` only
   after setting `disabled = NO` in the API clients.

In unrelated other fix, I return the JSON deserialization error
`JSONError` instead of the `NSURLSessionDataTask` error, which is `nil`,
if there is a JSON deserialization error.