Implement `setAndRememberClientId(,secret:)`, enabling complex init scenarios, especially with framework integrations
`WonderPush.setAndRememberClientId(,secret:)` not only initializes the
WonderPush SDK but also stores the given credentials so that next time
the SDK must initialize, it can do so on its own.

In complex customer scenarios, the credentials to be used cannot be
statically determined and are not available at startup, typically not
until the end-user selects a country or brand.

The WonderPush SDK however needs to be initialized in a few scenarios,
for instance when a push notification is clicked. When the SDK is
integrated through the use of some framework, it is not always easy or
feasible to have a synchronous communication with the code within the
framework at such time, so the native code can request the SDK to
initialize reusing remembered credentials.

Because the credentials are necessarily stored in a format readily
accessible for the application (although not to other applications),
it is necessary to opt into this behaviour by:

1. Calling `WonderPush.setClientId("USE_REMEMBERED", secret: "USE_REMEMBERED")`
   in the `application:willFinishLaunchingWithOptions:` method of your
   `AppDelegate`, instead of synchronously determining the appropriate
   values. You can alternatively use the value `nil`.
2. Call `WonderPush.setAndRememberClientId("…", secret: "…")` with the
   real Client ID and Client Secret when the appropriate values are
   known.

Providing `nil` values to `WonderPush.setAndRememberClientId(,secret:)` will
remove the previously stored values.